+++ src/config/types.rs
@@ -5,6 +5,15 @@ use std::str::FromStr;
 
 arg_enum! {
     #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Deserialize, Serialize)]
+    pub enum Color {
+        Auto,
+        Always,
+        Never,
+    }
+}
+
+arg_enum! {
+    #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Ord, PartialOrd, Deserialize, Serialize)]
     pub enum Mode {
         Test,
         Build
@@ -28,7 +37,6 @@ arg_enum! {
     #[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Deserialize, Serialize)]
     pub enum OutputFile {
         Json,
-        Toml,
         Stdout,
         Xml,
         Html,