+++ tests/workspaces.rs
@@ -2,6 +2,7 @@ use crate::utils::get_test_path;
 use cargo_tarpaulin::config::Config;
 use cargo_tarpaulin::launch_tarpaulin;
 use std::env;
+use std::path::PathBuf;
 
 #[test]
 fn package_exclude() {
@@ -63,4 +64,21 @@ fn specify_package() {
     let files = result.files();
     assert!(files.iter().any(|f| f.ends_with("foo/src/lib.rs")));
     assert!(files.iter().any(|f| f.ends_with("bar/src/lib.rs")));
+}
+
+#[test]
+fn config_relative_pathing() {
+    let mut test_dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR").unwrap());
+    test_dir.push("tests");
+    test_dir.push("data");
+    let base_path = test_dir.clone();
+    test_dir.push("tarpaulin.toml");
+    // This test added because if it doesn't work it can mess up using --features
+    // in workspace roots
+    let configs = Config::load_config_file(&test_dir).unwrap();
+
+    assert_eq!(configs.len(), 2);
+    assert_eq!(configs[0].manifest, base_path.join("lib/Cargo.toml"));
+    assert_eq!(configs[1].manifest, base_path.join("bin/Cargo.toml"));
+    assert_eq!(configs[1].target_dir(), base_path.join("targ"));
 }